useGasSharedValues
